Updates and Fixes
There were 2 device edit modals, we removed it to a single modal.
We added sidebar, rooms, devices and consumer table into the producer page.
[10:46 PM]
Added date of birth to consumer's settings.
Producer's api file is reserved for security and order.
Locking records of the doors are included in the log section.
We added mode selection to air conditioner and refrigerator.
The sensors have completely changed, a button has been added to turn the automatic data sending of the sensor on and off.
The responsive problem in the header on the landing page has been resolved.